Conflicts:
	include/xerus/ttNetwork.h
	src/xerus/contractionHeuristic.cpp
	src/xerus/misc/blasLapackWrapper.cpp
	src/xerus/misc/selectedFunctions.cpp
	src/xerus/ttNetwork.cpp

+++ include/xerus/contractionHeuristic.h
@@ -25,7 +25,7 @@ namespace xerus {
 
     //TODO individuelle schranken
     
-    /// Class representing used for heuristics to find good contraction orders of TensorNEtworks.
+    /// Class representing heuristics used to find good contraction orders of TensorNetworks.
     class ContractionHeuristic {
     public:
         static std::vector<ContractionHeuristic> *list;
@@ -38,15 +38,16 @@ namespace xerus {
             }
         };
         
+		/// Name of the heuristic
         std::string name;
         
-        float score;
+        double score;
         
         std::vector<std::pair<size_t,size_t>> contractions;
         
-        std::function<void(float &, std::vector<std::pair<size_t,size_t>> &, TensorNetwork &)> scoreFct;
+        std::function<void(double &, std::vector<std::pair<size_t,size_t>> &, TensorNetwork &)> scoreFct;
         
-        ContractionHeuristic(std::string _name, std::function<void(float &, std::vector<std::pair<size_t,size_t>> &, TensorNetwork &)> _scoreFct);
+        ContractionHeuristic(std::string _name, std::function<void(double &, std::vector<std::pair<size_t,size_t>> &, TensorNetwork &)> _scoreFct);
         
         double rescore(TensorNetwork _tn); // NOTE take as value to get a deep copy instead of reference!
     };
